[frontend] remove old report knowledge graph component
